requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
shazamio>=0.4.0
pydantic-settings>=2.0.0,<3.0.0
